feat(src) add AddContactScreen and detach Add Contact Logic from Home View
