+++ App.tsx
@@ -4,6 +4,7 @@ import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import HomeScreen from './src/views/Home';
 import DetailsScreen from './src/views/Details';
+import AddContactScreen from './src/views/AddContact';
 import {RootStackParamList} from './src/types/types';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
@@ -13,6 +14,7 @@ function RootStack() {
     <Stack.Navigator>
       <Stack.Screen name="Home" component={HomeScreen} />
       <Stack.Screen name="Details" component={DetailsScreen} />
+      <Stack.Screen name="AddContact" component={AddContactScreen} />
     </Stack.Navigator>
   );
 }
